actions: Split out zerver.lib.recipient_users.
